requirements.txt
absl-py==2.0.0
alembic==1.12.1
altair==5.2.0
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
Brotli==1.0.9
cachetools==5.3.2
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
cloudpickle==2.2.1
contourpy==1.2.0
cryptography==41.0.3
cucim==23.10.0
cycler==0.12.1
databricks-cli==0.18.0
docker==6.1.3
einops==0.6.1
entrypoints==0.4
filelock==3.13.1
fire==0.5.0
Flask==3.0.0
fonttools==4.45.1
fsspec==2023.10.0
gdown==4.7.1
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
google-auth-oauthlib==1.1.0
greenlet==3.0.1
grpcio==1.59.3
gunicorn==21.2.0
huggingface-hub==0.19.4
idna==3.4
imagecodecs==2023.9.18
imageio==2.33.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
itk==5.3.0
itk-core==5.3.0
itk-filtering==5.3.0
itk-io==5.3.0
itk-numerics==5.3.0
itk-registration==5.3.0
itk-segmentation==5.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kiwisolver==1.4.5
lazy_loader==0.3
lmdb==1.4.1
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl
mkl-fft
mkl-random
mkl-service
mlflow
monai==0.9.0
networkx==3.2.1
nibabel==5.1.0
nptyping==2.5.0
numpy==1.24.3
oauthlib==3.2.2
olefile==0.46
openslide-python==1.1.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pip==23.3.1
protobuf==4.23.4
psutil==5.9.6
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pynrrd==1.0.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-ignite==0.4.8
pytz==2023.3.post1
PyYAML==6.0.1
querystring-parser==1.2.4
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
sacremoses==0.1.1
safetensors==0.4.1
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
setuptools==68.0.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==2.0.23
sqlparse==0.4.4
streamlit==1.28.2
streamlit-cropper==0.2.2
streamlit-drawable-canvas==0.9.3
streamlit-image-coordinates==0.1.6
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.9.26
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tornado==6.4
tqdm==4.66.1
transformers==4.18.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.2
urllib3==1.26.18
validators==0.22.0
watchdog==3.0.0
websocket-client==1.6.4
Werkzeug==3.0.1
wheel==0.41.2
zipp==3.17.0
